FW/Random: change the return type of `generateInt()` to `long`
POSIX defines the `random()`[1] function to return `long`, in spite of
being limited to 2^31 - 1. By matching the return type, the compiler
will be able to tail-call into the virtual. Right now, it must place a
regular call so it can sign-extend to `long`.

Before:
```
00000000005bbdc0 <random>:
       endbr64
       sub    $0x8,%rsp
       mov    $0xfffffffffffffe30,%rax
       mov    0x6e8d37a(%rip),%rdi
:aligned_storage+0x148>
       mov    %fs:(%rax),%esi
       mov    (%rdi),%rax
       inc    %esi
       movslq %esi,%rsi
       shl    $0x6,%rsi
       add    0x8(%rdi),%rsi
       call   *0x50(%rax)
       add    $0x8,%rsp
       cltq
       ret
```

After:
```
<random>:
        endbr64
        mov    $0xfffffffffffffe30,%rax
        mov    0x6e8d37e(%rip),%rdi
        mov    %fs:(%rax),%esi
        mov    (%rdi),%rax
        inc    %esi
        movslq %esi,%rsi
        shl    $0x6,%rsi
        add    0x8(%rdi),%rsi
        jmp    *0x50(%rax)
```

Though this affected `rand()` and removed the tail-call it had, for no
good reason since the return register is unmodified.

[1] https://pubs.opengroup.org/onlinepubs/9799919799/functions/initstate.html

+++ framework/random.cpp
@@ -108,7 +108,7 @@ struct RandomEngineWrapper
     virtual uint32_t generate32(thread_rng *thread_buffer) = 0;
     virtual uint64_t generate48(thread_rng *thread_buffer) = 0;
     virtual uint64_t generate64(thread_rng *thread_buffer) = 0;
-    virtual int generateInt(thread_rng *thread_buffer) = 0;
+    virtual long int generateInt(thread_rng *thread_buffer) = 0;
     virtual __uint128_t generate128(thread_rng *thread_buffer) = 0;
 };
 RandomEngineWrapper::~RandomEngineWrapper() {}
@@ -195,7 +195,7 @@ template <typename E> struct EngineWrapper : public RandomEngineWrapper
         return generate32(thread_buffer) | (uint64_t(generate32(thread_buffer)) << 32);
     }
 
-    int generateInt(thread_rng *thread_buffer) override
+    long int generateInt(thread_rng *thread_buffer) override
     {
         return generate32(thread_buffer) & 0x7fffffff;
     }
@@ -289,7 +289,7 @@ uint64_t EngineWrapper<std::minstd_rand>::generate64(thread_rng *generator)
 }
 
 template <>
-int EngineWrapper<std::minstd_rand>::generateInt(thread_rng *generator)
+long int EngineWrapper<std::minstd_rand>::generateInt(thread_rng *generator)
 {
     // need a single sample to make positive int
     return engine(generator)();
